% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsampling-package.R
\docType{package}
\name{emsampling-package}
\alias{emsampling}
\alias{emsampling-package}
\title{emsampling: Experimental functions to analyse EM connectomics data}
\description{
Experimental functions to analyse the results of electron 
    microscopy connectomics data, focussed on the common situation where only
    some of the connections for a given neuron have been identified.
}
\section{Getting started}{
 It is worth looking at the documentation main class
  in the package, \code{\link{samplingcurve}}, as well as the articles on the
  \href{https://jefferis.github.io/emsampling}{documentation website}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://jefferis.github.io/emsampling}
  \item \url{https://github.com/jefferis/emsampling}
  \item Report bugs at \url{https://github.com/jefferis/emsampling/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

}
\keyword{internal}
