% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_curves.R
\name{hist.samplingcurve}
\alias{hist.samplingcurve}
\title{Plot a histogram of connections per partner neuron}
\usage{
\method{hist}{samplingcurve}(x, decreasing = TRUE, plot = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link{samplingcurve}} object}

\item{...}{}
}
\description{
Strictly speaking this is a bar plot for a table object rather
  than an R histogram
}
