% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pd2a1.1.sc}
\alias{pd2a1.1.sc}
\title{samplingcurve object for dendritic inputs of PD2a1#1 lateral horn neuron}
\format{An object of class \code{samplingcurve} (inherits from \code{data.frame}) with 1035 rows and 2 columns.}
\usage{
pd2a1.1.sc
}
\description{
samplingcurve object for dendritic inputs of PD2a1#1 lateral horn neuron
}
\details{
See \code{data-raw/pd2a1.1.R} for details of how this object was
  constructed.
}
\examples{
plot(hist(pd2a1.1.sc))
}
\references{
See \url{https://www.biorxiv.org/content/early/2017/08/29/167312}
}
\keyword{datasets}
