% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_curves.R
\name{plot.samplingcurve}
\alias{plot.samplingcurve}
\title{Plot a standard sampling surve}
\usage{
\method{plot}{samplingcurve}(x, col = "red", ...)
}
\arguments{
\item{...}{}
}
\description{
Plot a standard sampling surve
}
