% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_curves.R
\name{samplingcurve}
\alias{samplingcurve}
\alias{plot.samplingcurve}
\alias{lines.samplingcurve}
\alias{hist.samplingcurve}
\title{Make a basic sampling curve from a vector of partner ids}
\usage{
samplingcurve(partners, N = NULL, m = NULL)

\method{plot}{samplingcurve}(x, col = "red", ...)

\method{lines}{samplingcurve}(x, rand = 0, mean = FALSE, lty = 3,
  col = NULL, ..., colpal = "grey")

\method{hist}{samplingcurve}(x, decreasing = TRUE, plot = TRUE, ...)
}
\arguments{
\item{partners}{A vector or partner neuron identifiers (typically numeric
such as CATMAID skeleton ids)}

\item{N, m}{optional parameters describing the total number of connections and
the total number of partners (if known).}

\item{x}{A \code{\link{samplingcurve}} object}

\item{col}{line colour (see \code{\link{lines}})}

\item{...}{Additional arguments to plotting functions}

\item{rand}{number of randomised versions of curve to plot}

\item{mean}{whether to plot the mean of specified number of random curves
rather than each individual curve.}

\item{lty}{line type (see \code{\link{par}})}

\item{colpal}{A colour palette. Either a function (see
\code{\link[grDevices]{rainbow}}) or a vector of colour names. Only used
when \code{col=NULL})}

\item{decreasing}{Whether to plot the strongest connections closest to the y
axis (default \code{TRUE})}

\item{plot}{Whether to show the histogram}
}
\value{
An object of class \code{samplingcurve}, currently implemented as a
  data.frame.

\code{hist.samplingcurve} returns the \code{table} of connections per
  partner used for the plot.
}
\description{
\code{plot.samplingcurve} plots a standard sampling curve

\code{lines.samplingcurve} adds a line for a sampling curve to
  an existing plot. It can also add a specified number of re-randomised
  versions of the curve, optionally producing a smoothed mean.

\code{hist.samplingcurve} plots a histogram of connections per
  partner neuron. strictly speaking this is a bar plot for a table object
  rather than an R histogram
}
\examples{
scuniform=samplingcurve(sample(1:20, size=200, replace=TRUE))
plot(scuniform)
# add 20 random realisations
lines(scuniform, rand=20)
# add a smooth mean
lines(scuniform, rand=1000, mean=TRUE, col='black')

# use real sample data for inputs to a lateral horn neuron
plot(pd2a1.1.sc)
lines(pd2a1.1.sc, rand=20)
scuniform=samplingcurve(sample(1:20, size=200, replace=TRUE))
hist(scuniform, main='20 neurons with equal connection probability')

hist(pd2a1.1.sc, main='Inputs to a lateral horn neuron')
}
