% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_curves.R
\name{subsample}
\alias{subsample}
\title{Reorder a sampling curve object, optionally subsampling a given fraction}
\usage{
subsample(x, fraction = 1)
}
\arguments{
\item{fraction}{}
}
\description{
Reorder a sampling curve object, optionally subsampling a given fraction
}
