% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_curves.R
\name{urmvhyper}
\alias{urmvhyper}
\title{Generate a set of replicates for uniform multivate hypergeometric distribution}
\usage{
urmvhyper(N, m, k = NULL, fraction = NULL, nn = 10000)
}
\arguments{
\item{N}{number of marbles}

\item{m}{number of colours}

\item{k}{sample size}

\item{fraction}{Fraction of N to sample}

\item{nn}{Number of replicates}
}
\description{
Generate a set of replicates for uniform multivate hypergeometric distribution
}
