% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsampling-package.R
\docType{package}
\name{emsampling-package}
\alias{emsampling}
\alias{emsampling-package}
\title{emsampling: Experimental functions to analyse EM connectomics data}
\description{
Experimental functions to analyse the results of electron 
    microscopy connectomics data, focussed on the common situation where only
    some of the connections for a given neuron have been identified.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jefferis.github.io/emsampling}
  \item \url{https://github.com/jefferis/emsampling}
  \item Report bugs at \url{https://github.com/jefferis/emsampling/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

}
\keyword{It}
\keyword{\code{\link{samplingcurve}},}
\keyword{articles}
\keyword{as}
\keyword{at}
\keyword{class}
\keyword{documentation}
\keyword{in}
\keyword{internal}
\keyword{is}
\keyword{looking}
\keyword{main}
\keyword{on}
\keyword{package,}
\keyword{the}
\keyword{website.}
\keyword{well}
\keyword{worth}
