% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_curves.R
\name{make_rand_sampling_curve}
\alias{make_rand_sampling_curve}
\title{Make (randomised) sampling curve object for neuron or connection table}
\usage{
make_rand_sampling_curve(x, partners = c("auto", "out", "in"),
  sample = TRUE)
}
\arguments{
\item{x}{A \code{\link{catmaid_skids}} compatible neuron specification or a
data.frame generated by \code{\link{catmaid_get_connector_table}}.}

\item{partners}{Whether to examine the input or output partners for given
neuron.}

\item{sample}{Whether to randomise the connection order (this is the right
thing to do if the input neuron has been completed but was not randomly
sampled in the first place.)}
}
\value{
An object of class \code{\link{samplingcurve}}
}
\description{
Make (randomised) sampling curve object for neuron or connection table
}
\details{
\code{make_rand_sampling_curve} is a thin wrapper for
  \code{\link{samplingcurve}}, which has the additional functionality of
  fetching (all) the synapses for one or more CATMAID neurons (specified
  using any form compatible with \code{\link{catmaid_skids}}).
}
\examples{
\donttest{
sc=make_rand_sampling_curve(2560371)
head(sc)
plot(sc)
}
}
