% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_curves.R
\name{samplingcurve}
\alias{samplingcurve}
\title{Make a basic sampling curve from a vector of partner ids}
\usage{
samplingcurve(partners, N = NULL, m = NULL, ...)
}
\arguments{
\item{N, m}{optional parameters describing the total number of connections and
the total number of partners (if known).}

\item{...}{}
}
\value{
An object of class \code{samplingcurve}, currently implemented as a
  data.frame.
}
\description{
Make a basic sampling curve from a vector of partner ids
}
\examples{
scuniform=samplingcurve(sample(1:20, size=200, replace=T))
plot(scuniform)
}
