% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_curves.R
\name{subsample}
\alias{subsample}
\title{Reorder a sampling curve object, optionally subsampling a given fraction}
\usage{
subsample(x, fraction = 1)
}
\arguments{
\item{x}{A \code{samplingcurve} object or a neuron specification that can be
passed to \code{make_rand_sampling_curve}.}

\item{fraction}{A fraction from 0-1 speciying the proportion of connections
to keep in the sample.}
}
\value{
A new \code{samplingcurve} object
}
\description{
Reorder a sampling curve object, optionally subsampling a given fraction
}
\examples{
scuniform=samplingcurve(sample(1:20, size=200, replace=TRUE))
hist(scuniform)
head(scuniform)
}
